feat(api): add project, task file, and task services for management
